+++ pydses/__init__.py
@@ -1,0 +1,6 @@
+"""
+@author: grimrhapsody
+"""
+
+from .event_writer import *
+from .event_enums import *
